refactor: extract subview components from large view files
Extract 6 struct views from the 3 largest view files to create proper SwiftUI invalidation boundaries:

AgentChatView (1321 → 551 lines):
- AgentChatHeaderView: status dot, model picker, thinking level menu
- AgentPromptView: prompt capsule, cursor, action button, completion popup
- AgentToolCallCard: tool call display with status, result, duration
- AgentApprovalCard: approval request with diff preview and action buttons

TabBarView (840 → 713 lines):
- WorkspaceIndicatorView: rename field, icon picker, context menu
- AgentStatusDot: shared status indicator (also used by collapsed capsule)

ContentView (922 → 885 lines):
- ChangeSummarySidebarView: resizable sidebar with drag handle
