+++ macos/Sources/Views/AgentChatView.swift
@@ -28,9 +28,6 @@ struct AgentChatView: View {
     /// Tracks whether the user has scrolled away from the bottom.
     /// When true, auto-scroll is paused to let the user read earlier content.
     @State private var userHasScrolledUp: Bool = false
-    @State private var isModelHovered: Bool = false
-    @State private var isThinkingHovered: Bool = false
-    @State private var isHelpHovered: Bool = false
     @Environment(\.accessibilityReduceMotion) private var reduceMotion
     /// Whether auto-scroll should follow streaming output.
     private var shouldAutoScroll: Bool { !userHasScrolledUp }
@@ -38,7 +35,7 @@ struct AgentChatView: View {
     var body: some View {
         VStack(spacing: 0) {
             // Header bar
-            chatHeader
+            AgentChatHeaderView(state: state, encoder: encoder)
 
             // Messages: bottom-anchored so few messages cluster near the
             // prompt input rather than leaving a void below them.
@@ -115,169 +112,10 @@ struct AgentChatView: View {
                 }
             }
 
-            // Prompt completion popup (floats above the prompt area)
-            if let completion = state.promptCompletion {
-                promptCompletionPopup(completion)
-            }
-
-            // Prompt area
-            promptArea
+            // Prompt (completion popup + input capsule)
+            AgentPromptView(state: state, isInsertMode: isInsertMode, encoder: encoder)
         }
         .background(theme.agentPanelBg)
-    }
-
-    // MARK: - Header
-
-    @ViewBuilder
-    private var chatHeader: some View {
-        HStack(spacing: 8) {
-            // Status indicator
-            Circle()
-                .fill(statusColor)
-                .frame(width: 7, height: 7)
-                .accessibilityLabel("Agent status: \(state.statusLabel)")
-
-            modelPickerButton
-
-            Text("·")
-                .font(.system(size: 11))
-                .foregroundStyle(theme.agentDisabledFg)
-
-            thinkingLevelMenu
-
-            if state.isThinking {
-                ProgressView()
-                    .controlSize(.small)
-                    .scaleEffect(0.6)
-            }
-
-            Spacer()
-
-            Text(state.statusLabel)
-                .font(.system(size: 11))
-                .foregroundStyle(theme.agentMutedFg)
-
-            Button {
-                // Send '?' to toggle help overlay
-                encoder?.sendKeyPress(codepoint: 0x3F, modifiers: 0)
-            } label: {
-                Image(systemName: "questionmark.circle")
-                    .font(.system(size: 13))
-                    .foregroundStyle(state.helpVisible ? theme.agentHeaderFg : theme.agentMutedFg)
-                    .padding(.horizontal, 4)
-                    .padding(.vertical, 3)
-                    .background(RoundedRectangle(cornerRadius: 4).fill(isHelpHovered ? theme.agentTextFg.opacity(0.06) : Color.clear))
-            }
-            .buttonStyle(.plain)
-            .help("Keyboard shortcuts (?)")
-            .accessibilityLabel("Agent help")
-            .accessibilityHint("Shows keybinding cheatsheet")
-            .accessibilityAddTraits(.isButton)
-            .onHover { hovering in
-                isHelpHovered = hovering
-            }
-            .pointingHandCursor()
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 8)
-        .background(theme.agentHeaderBg)
-
-        Rectangle()
-            .fill(theme.agentCodeBorder.opacity(0.3))
-            .frame(height: 1)
-    }
-
-    private var modelPickerButton: some View {
-        Button {
-            if !state.isThinking {
-                encoder?.sendExecuteCommand(name: "agent_pick_model")
-            }
-        } label: {
-            HStack(spacing: 4) {
-                Text(state.displayModel.isEmpty ? "Agent" : state.displayModel)
-                    .font(.system(size: 12, weight: .medium))
-                Image(systemName: "chevron.up.chevron.down")
-                    .font(.system(size: 9, weight: .semibold))
-            }
-            .foregroundStyle(state.isThinking ? theme.agentSecondaryFg : theme.agentHeaderFg)
-            .padding(.horizontal, 6)
-            .padding(.vertical, 3)
-            .background(RoundedRectangle(cornerRadius: 4).fill(isModelHovered && !state.isThinking ? theme.agentTextFg.opacity(0.06) : Color.clear))
-        }
-        .buttonStyle(.plain)
-        .allowsHitTesting(!state.isThinking)
-        .help("Pick model (SPC a m)")
-        .accessibilityLabel("Agent model")
-        .accessibilityValue(state.displayModel.isEmpty ? "Agent" : state.displayModel)
-        .accessibilityHint(state.isThinking ? "Disabled while the agent is streaming" : "Opens the model picker")
-        .accessibilityAddTraits(.isButton)
-        .onHover { isModelHovered = $0 }
-        .pointingHandCursor(isEnabled: !state.isThinking)
-    }
-
-    @ViewBuilder
-    private var thinkingLevelMenu: some View {
-        if state.isThinking {
-            thinkingLevelLabel
-                .help("Pick thinking level (SPC a T)")
-                .accessibilityLabel("Agent thinking level")
-                .accessibilityValue(state.thinkingLabel)
-                .accessibilityHint("Disabled while the agent is streaming")
-                .accessibilityAddTraits(.isButton)
-        } else {
-            Menu {
-                ForEach(["off", "low", "medium", "high"], id: \.self) { level in
-                    Button {
-                        encoder?.sendExecuteCommand(name: "agent_thinking_\(level)")
-                    } label: {
-                        HStack {
-                            Text(thinkingDisplayName(level))
-                            if state.thinkingLevel == level {
-                                Image(systemName: "checkmark")
-                            }
-                        }
-                    }
-                    .accessibilityLabel("Set thinking level to \(thinkingDisplayName(level))")
-                    .accessibilityHint("Changes the agent thinking level to \(thinkingDisplayName(level))")
-                }
-            } label: {
-                thinkingLevelLabel
-            }
-            .menuStyle(.borderlessButton)
-            .fixedSize()
-            .help("Pick thinking level (SPC a T)")
-            .accessibilityLabel("Agent thinking level")
-            .accessibilityValue(state.thinkingLabel)
-            .accessibilityHint("Opens a menu of thinking levels")
-            .accessibilityAddTraits(.isButton)
-            .onHover { isThinkingHovered = $0 }
-            .pointingHandCursor()
-        }
-    }
-
-    private var thinkingLevelLabel: some View {
-        HStack(spacing: 4) {
-            Image(systemName: state.thinkingIconName)
-                .font(.system(size: 12))
-            Text(state.thinkingLabel)
-                .font(.system(size: 12, weight: .medium))
-            Image(systemName: "chevron.down")
-                .font(.system(size: 8, weight: .semibold))
-        }
-        .foregroundStyle(state.isThinking ? theme.agentSecondaryFg : theme.agentHeaderFg.opacity(0.9))
-        .padding(.horizontal, 6)
-        .padding(.vertical, 3)
-        .background(RoundedRectangle(cornerRadius: 4).fill(isThinkingHovered && !state.isThinking ? theme.agentTextFg.opacity(0.06) : Color.clear))
-    }
-
-    private func thinkingDisplayName(_ level: String) -> String {
-        switch level {
-        case "off": return "Off"
-        case "low": return "Low"
-        case "medium": return "Medium"
-        case "high": return "High"
-        default: return level.capitalized
-        }
     }
 
     // MARK: - Messages
@@ -311,12 +149,16 @@ struct AgentChatView: View {
             assistantMarkdownBlock(blocks)
         case .thinking(_, let text, let collapsed):
             thinkingBlock(text, collapsed: collapsed)
-        case .toolCall(_, let name, let summary, let status, let isError, let collapsed, let autoApprovedScope, let duration, let result, let previewKind, let previewLines):
-            toolCallCard(messageIndex: index, name: name, summary: summary, status: status, isError: isError, collapsed: collapsed, autoApprovedScope: autoApprovedScope, durationMs: duration, result: result, resultLines: nil, previewKind: previewKind, previewLines: previewLines)
-        case .styledToolCall(_, let name, let summary, let status, let isError, let collapsed, let autoApprovedScope, let duration, let resultLines, let previewKind, let previewLines):
-            toolCallCard(messageIndex: index, name: name, summary: summary, status: status, isError: isError, collapsed: collapsed, autoApprovedScope: autoApprovedScope, durationMs: duration, result: nil, resultLines: resultLines, previewKind: previewKind, previewLines: previewLines)
+        case .toolCall(_, let name, let summary, let status, let isError, let collapsed, let autoApprovedScope, let duration, let result, _, let previewLines):
+            AgentToolCallCard(messageIndex: index, name: name, summary: summary, status: status, isError: isError, collapsed: collapsed, autoApprovedScope: autoApprovedScope, durationMs: duration, result: result, resultLines: nil, previewLines: previewLines, encoder: encoder, styledLineView: { runs, fontSize, mono in
+                AnyView(styledLineView(runs, baseFontSize: fontSize, monospaced: mono))
+            })
+        case .styledToolCall(_, let name, let summary, let status, let isError, let collapsed, let autoApprovedScope, let duration, let resultLines, _, let previewLines):
+            AgentToolCallCard(messageIndex: index, name: name, summary: summary, status: status, isError: isError, collapsed: collapsed, autoApprovedScope: autoApprovedScope, durationMs: duration, result: nil, resultLines: resultLines, previewLines: previewLines, encoder: encoder, styledLineView: { runs, fontSize, mono in
+                AnyView(styledLineView(runs, baseFontSize: fontSize, monospaced: mono))
+            })
         case .approvalToolCall(_, let name, let summary, let toolCallId, let previewKind, let previewLines):
-            approvalToolCallCard(name: name, summary: summary, toolCallId: toolCallId, previewKind: previewKind, previewLines: previewLines)
+            AgentApprovalCard(name: name, summary: summary, toolCallId: toolCallId, previewKind: previewKind, previewLines: previewLines, encoder: encoder)
         case .system(_, let text, let isError):
             systemMessage(text, isError: isError)
         case .usage(_, let input, let output, _, _, let costMicros):
@@ -575,113 +417,6 @@ struct AgentChatView: View {
     }
 
     @ViewBuilder
-    private func toolCallCard(messageIndex: Int, name: String, summary: String, status: UInt8, isError: Bool, collapsed: Bool, autoApprovedScope: UInt8, durationMs: UInt32, result: String?, resultLines: [[Wire.StyledTextRun]]?, previewKind _: UInt8, previewLines: [String]) -> some View {
-        let hasResult = result?.isEmpty == false || resultLines?.isEmpty == false
-        let visiblePreviewLines = Array(previewLines.prefix(8))
-
-        VStack(alignment: .leading, spacing: 0) {
-            // Header (clickable to toggle collapse)
-            HStack(spacing: 6) {
-                // Collapse/expand chevron (only shown when there's result content)
-                if hasResult {
-                    Image(systemName: collapsed ? "chevron.right" : "chevron.down")
-                        .font(.system(size: 9, weight: .medium))
-                        .foregroundStyle(theme.agentTextFg.opacity(0.4))
-                }
-
-                // Running spinner or status icon
-                if status == 0 {
-                    ProgressView()
-                        .controlSize(.mini)
-                        .scaleEffect(0.6)
-                } else {
-                    Image(systemName: toolIcon(status))
-                        .font(.system(size: 10))
-                        .foregroundStyle(isError ? Color.red.opacity(0.8) : theme.agentToolHeader)
-                }
-
-                Text(name)
-                    .font(.system(size: 12, weight: .medium, design: .monospaced))
-                    .foregroundStyle(theme.agentTextFg)
-                    .layoutPriority(1)
-
-                // Tool summary (command, path, etc.)
-                if !summary.isEmpty {
-                    toolCallSummaryView(name: name, summary: summary)
-                }
-
-                Spacer(minLength: 8)
-
-                if durationMs > 0 {
-                    Text(formatDuration(durationMs))
-                        .font(.system(size: 10))
-                        .foregroundStyle(theme.agentTextFg.opacity(0.3))
-                }
-
-                statusBadge(status, isError: isError)
-
-                if let autoApprovedLabel = autoApprovedScopeLabel(autoApprovedScope) {
-                    autoApprovedPill(label: autoApprovedLabel)
-                }
-            }
-            .padding(.horizontal, 10)
-            .padding(.vertical, 6)
-            .contentShape(Rectangle())
-            .onTapGesture {
-                if hasResult, messageIndex <= Int(UInt16.max) {
-                    encoder?.sendAgentToolToggle(index: UInt16(messageIndex))
-                }
-            }
-
-            if !visiblePreviewLines.isEmpty {
-                Rectangle()
-                    .fill(theme.agentToolBorder.opacity(0.2))
-                    .frame(height: 1)
-
-                approvalPreviewLines(visiblePreviewLines)
-                    .padding(.horizontal, 8)
-                    .padding(.bottom, 8)
-            }
-
-            // Result (collapsed by default, supports styled or plain text)
-            if !collapsed && hasResult {
-                Rectangle()
-                    .fill(theme.agentToolBorder.opacity(0.2))
-                    .frame(height: 1)
-
-                ScrollView([.horizontal, .vertical]) {
-                    if let lines = resultLines, !lines.isEmpty {
-                        // Styled result with tree-sitter/markdown formatting
-                        VStack(alignment: .leading, spacing: 2) {
-                            ForEach(Array(lines.enumerated()), id: \.offset) { _, runs in
-                                styledLineView(runs, baseFontSize: 11, monospaced: true)
-                            }
-                        }
-                        .padding(10)
-                    } else if let text = result, !text.isEmpty {
-                        // Plain text fallback
-                        Text(text)
-                            .font(.system(size: 11, design: .monospaced))
-                            .foregroundStyle(theme.agentTextFg.opacity(0.7))
-                            .textSelection(.enabled)
-                            .fixedSize(horizontal: true, vertical: false)
-                            .padding(10)
-                    }
-                }
-                .frame(maxHeight: 200)
-            }
-        }
-        .background(
-            RoundedRectangle(cornerRadius: 8)
-                .fill(theme.agentCodeBg)
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 8)
-                .strokeBorder(isError ? Color.red.opacity(0.3) : theme.agentToolBorder.opacity(0.3), lineWidth: 1)
-        )
-    }
-
-    @ViewBuilder
     private func systemMessage(_ text: String, isError: Bool) -> some View {
         HStack(spacing: 6) {
             Image(systemName: isError ? "exclamationmark.triangle" : "info.circle")
@@ -705,407 +440,6 @@ struct AgentChatView: View {
         }
         .font(.system(size: 10))
         .foregroundStyle(theme.agentTextFg.opacity(0.3))
-    }
-
-    @ViewBuilder
-    private func approvalToolCallCard(name: String, summary: String, toolCallId: String, previewKind: UInt8, previewLines: [String]) -> some View {
-        let visiblePreviewLines = Array(previewLines.prefix(8))
-
-        VStack(alignment: .leading, spacing: 10) {
-            approvalToolCallHeader(name: name, summary: summary, previewKind: previewKind)
-
-            if !visiblePreviewLines.isEmpty {
-                approvalPreviewLines(visiblePreviewLines)
-            }
-
-            approvalButtons(toolCallId: toolCallId)
-        }
-        .padding(12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(approvalCardBackground)
-        .overlay(approvalCardBorder)
-    }
-
-    private func approvalToolCallHeader(name: String, summary: String, previewKind: UInt8) -> some View {
-        HStack(alignment: .firstTextBaseline, spacing: 8) {
-            Image(systemName: "exclamationmark.shield")
-                .font(.system(size: 13))
-                .foregroundStyle(Color.orange)
-
-            Text("Approval required")
-                .font(.system(size: 12, weight: .semibold))
-                .foregroundStyle(theme.agentTextFg)
-
-            Text(name)
-                .font(.system(size: 12, weight: .bold, design: .monospaced))
-                .foregroundStyle(theme.agentToolHeader)
-
-            if !summary.isEmpty {
-                approvalToolCallSummary(summary: summary, previewKind: previewKind)
-            }
-        }
-    }
-
-    @ViewBuilder
-    private func toolCallSummaryView(name: String, summary: String) -> some View {
-        if name == "shell" {
-            ScrollView(.horizontal, showsIndicators: false) {
-                Text(summary)
-                    .font(.system(size: 11, design: .monospaced))
-                    .foregroundStyle(theme.agentTextFg.opacity(0.5))
-                    .lineLimit(1)
-                    .fixedSize(horizontal: true, vertical: false)
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-        } else {
-            Text(summary)
-                .font(.system(size: 11, design: .monospaced))
-                .foregroundStyle(theme.agentTextFg.opacity(0.5))
-                .lineLimit(1)
-                .truncationMode(.middle)
-        }
-    }
-
-    @ViewBuilder
-    private func approvalToolCallSummary(summary: String, previewKind: UInt8) -> some View {
-        if previewKind == 2 {
-            ScrollView(.horizontal, showsIndicators: false) {
-                Text(summary)
-                    .font(.system(size: 12, weight: .regular, design: .monospaced))
-                    .foregroundStyle(theme.agentTextFg)
-                    .lineLimit(1)
-                    .fixedSize(horizontal: true, vertical: false)
-            }
-        } else {
-            Text(summary)
-                .font(.system(size: 12, weight: .regular, design: .monospaced))
-                .foregroundStyle(theme.agentTextFg)
-                .lineLimit(2)
-        }
-    }
-
-    private func approvalPreviewLines(_ lines: [String]) -> some View {
-        VStack(alignment: .leading, spacing: 2) {
-            ForEach(Array(lines.enumerated()), id: \.offset) { _, line in
-                approvalPreviewLine(line)
-            }
-        }
-        .padding(8)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(approvalPreviewBackground)
-        .overlay(approvalPreviewBorder)
-    }
-
-    private func approvalPreviewLine(_ line: String) -> some View {
-        Text(line)
-            .font(.system(size: 10, design: .monospaced))
-            .foregroundStyle(approvalPreviewLineColor(line))
-            .truncationMode(.tail)
-    }
-
-    private func approvalPreviewLineColor(_ line: String) -> Color {
-        if line.hasPrefix("+") {
-            return Color.green.opacity(0.75)
-        }
-        if line.hasPrefix("-") {
-            return Color.red.opacity(0.75)
-        }
-        return theme.agentTextFg.opacity(0.65)
-    }
-
-    private func approvalButtons(toolCallId _: String) -> some View {
-        HStack(spacing: 6) {
-            Button("Approve (y)") {
-                encoder?.sendKeyPress(codepoint: 0x79, modifiers: 0)
-            }
-            .buttonStyle(.borderedProminent)
-            .controlSize(.small)
-
-            HStack(spacing: 6) {
-                Button("Trust for session (a)") {
-                    encoder?.sendKeyPress(codepoint: 0x61, modifiers: 0)
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-
-                Button("Trust for this turn (t)") {
-                    encoder?.sendKeyPress(codepoint: 0x74, modifiers: 0)
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-            }
-
-            Button("Deny (n)") {
-                encoder?.sendKeyPress(codepoint: 0x6E, modifiers: 0)
-            }
-            .buttonStyle(.bordered)
-            .controlSize(.small)
-        }
-    }
-
-    private var approvalCardBackground: some View {
-        RoundedRectangle(cornerRadius: 10).fill(theme.agentCodeBg.opacity(0.9))
-    }
-
-    private var approvalCardBorder: some View {
-        RoundedRectangle(cornerRadius: 10).stroke(Color.orange.opacity(0.5), lineWidth: 1)
-    }
-
-    private var approvalPreviewBackground: some View {
-        RoundedRectangle(cornerRadius: 6).fill(theme.agentCodeBg.opacity(0.8))
-    }
-
-    private var approvalPreviewBorder: some View {
-        RoundedRectangle(cornerRadius: 6).stroke(theme.agentCodeBorder.opacity(0.35), lineWidth: 1)
-    }
-
-    private func approvalPreviewLabel(_ kind: UInt8) -> String {
-        switch kind {
-        case 1: return "Diff"
-        case 2: return "Command"
-        case 3: return "Target"
-        default: return "Args"
-        }
-    }
-
-    // MARK: - Prompt
-
-    // MARK: - Prompt area
-
-    /// Whether the agent is actively streaming a response.
-    private var isStreaming: Bool { state.status == 1 || state.status == 2 }
-
-    /// Whether the send button should be enabled (insert mode with text).
-    private var canSend: Bool { isInsertMode && !state.prompt.isEmpty && !isStreaming }
-
-    /// The SF Symbol name for the action button, morphing between send and stop.
-    private var actionButtonIcon: String {
-        isStreaming ? "stop.circle.fill" : "arrow.up.circle.fill"
-    }
-
-    /// The action button's foreground color based on state.
-    private var actionButtonColor: Color {
-        if isStreaming { return .red }
-        if canSend { return theme.agentInputBorder }
-        return theme.agentTextFg.opacity(0.2)
-    }
-
-    /// Capsule border color: accent when in insert mode, subtle border otherwise.
-    private var capsuleBorderColor: Color {
-        if isInsertMode { return theme.agentInputBorder.opacity(0.5) }
-        return theme.agentCodeBorder.opacity(0.3)
-    }
-
-    /// Capsule background opacity shifts with mode.
-    private var capsuleBgOpacity: Double {
-        if isInsertMode { return 0.8 }
-        if isStreaming { return 0.6 }
-        return 0.4
-    }
-
-    /// Vim mode label shown in the prompt border.
-    private var modeLabel: String {
-        switch state.promptVimMode {
-        case 0: return "NORMAL"
-        case 2: return "VISUAL"
-        case 3: return "V-LINE"
-        case 4: return "OP"
-        default: return "" // insert mode: no label
-        }
-    }
-
-    @ViewBuilder
-    private var promptArea: some View {
-        HStack(spacing: 8) {
-            promptCapsule
-            actionButton
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 8)
-        .accessibilityElement(children: .contain)
-        .accessibilityLabel("Agent prompt")
-    }
-
-    @ViewBuilder
-    private var promptCapsule: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            // Mode indicator bar (only in non-insert modes)
-            if !modeLabel.isEmpty {
-                HStack(spacing: 4) {
-                    Text(modeLabel)
-                        .font(.system(size: 9, weight: .bold, design: .monospaced))
-                        .foregroundStyle(theme.agentInputBorder)
-                    Spacer()
-                }
-                .padding(.horizontal, 12)
-                .padding(.top, 4)
-                .padding(.bottom, 2)
-            }
-
-            // Prompt text with monospaced font and cursor
-            HStack(spacing: 0) {
-                if isStreaming && state.prompt.isEmpty {
-                    Text("Generating...")
-                        .font(.system(size: 13, design: .monospaced))
-                        .foregroundStyle(theme.agentInputPlaceholder)
-                        .italic()
-                } else if state.prompt.isEmpty && !isInsertMode {
-                    Text("Ask anything...")
-                        .font(.system(size: 13, design: .monospaced))
-                        .foregroundStyle(theme.agentInputPlaceholder)
-                } else {
-                    // Render prompt text with cursor overlay
-                    promptTextWithCursor
-                }
-
-                Spacer(minLength: 4)
-            }
-            .padding(.horizontal, 12)
-            .padding(.vertical, modeLabel.isEmpty ? 10 : 6)
-        }
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(theme.agentInputBg.opacity(capsuleBgOpacity))
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 12)
-                .strokeBorder(capsuleBorderColor, lineWidth: 1)
-        )
-        .contentShape(RoundedRectangle(cornerRadius: 12))
-        .onTapGesture {
-            if !isInsertMode && !isStreaming {
-                encoder?.sendKeyPress(codepoint: 0x69, modifiers: 0)
-            }
-        }
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel("Chat input")
-        .accessibilityValue(state.prompt.isEmpty ? "Empty" : state.prompt)
-        .accessibilityHint(isInsertMode ? "Type a message, press Return to send" : "Press i to start typing")
-    }
-
-    /// Monospace character width at the prompt font size, computed from actual font metrics.
-    private var promptCharWidth: CGFloat {
-        let font = NSFont.monospacedSystemFont(ofSize: 13, weight: .regular)
-        let size = ("M" as NSString).size(withAttributes: [.font: font])
-        return size.width
-    }
-
-    /// Line height for the prompt font, derived from actual font metrics.
-    private var promptLineHeight: CGFloat {
-        let font = NSFont.monospacedSystemFont(ofSize: 13, weight: .regular)
-        return ceil(font.ascender - font.descender + font.leading)
-    }
-
-    /// Renders the prompt text with a cursor at the BEAM-reported position.
-    /// Uses monospaced font so cursor positioning aligns with character columns.
-    @ViewBuilder
-    private var promptTextWithCursor: some View {
-        let lines = state.prompt.split(separator: "\n", omittingEmptySubsequences: false).map(String.init)
-        let cursorLine = Int(state.promptCursorLine)
-        let cursorCol = Int(state.promptCursorCol)
-        let isBlock = state.promptVimMode == 0 || state.promptVimMode >= 2
-        let charW = promptCharWidth
-        let lineH = promptLineHeight
-
-        VStack(alignment: .leading, spacing: 2) {
-            ForEach(Array(lines.prefix(8).enumerated()), id: \.offset) { lineIdx, line in
-                ZStack(alignment: .leading) {
-                    Text(line.isEmpty ? " " : line)
-                        .font(.system(size: 13, design: .monospaced))
-                        .foregroundStyle(theme.agentTextFg.opacity(isStreaming ? 0.4 : 1.0))
-
-                    if lineIdx == cursorLine && !isStreaming {
-                        let cursorX = CGFloat(cursorCol) * charW
-
-                        if isBlock {
-                            Rectangle()
-                                .fill(theme.agentInputBorder.opacity(0.7))
-                                .frame(width: charW, height: lineH)
-                                .offset(x: cursorX)
-                        } else {
-                            Rectangle()
-                                .fill(theme.agentInputBorder)
-                                .frame(width: 1.5, height: lineH)
-                                .offset(x: cursorX)
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var actionButton: some View {
-        Button {
-            if isStreaming {
-                // Send Ctrl+C to abort
-                encoder?.sendKeyPress(codepoint: 0x63, modifiers: 0x02)
-            } else if canSend {
-                // Send Enter to submit
-                encoder?.sendKeyPress(codepoint: 0x0D, modifiers: 0)
-            }
-        } label: {
-            Image(systemName: actionButtonIcon)
-                .font(.system(size: 24))
-                .foregroundStyle(actionButtonColor)
-                .contentTransition(.symbolEffect(.replace))
-        }
-        .buttonStyle(.plain)
-        .disabled(!canSend && !isStreaming)
-        .accessibilityLabel(isStreaming ? "Stop generating" : "Send message")
-        .accessibilityHint(isStreaming ? "Sends Ctrl+C to abort" : "Sends the current prompt")
-        .accessibilityAddTraits(.isButton)
-    }
-
-    // MARK: - Prompt completion popup
-
-    /// Renders a floating completion popup for @-mention or /slash commands.
-    /// Styled to match the existing CompletionOverlay used for LSP completions.
-    @ViewBuilder
-    private func promptCompletionPopup(_ completion: Wire.PromptCompletion) -> some View {
-        let isSlash = completion.type == 1
-
-        VStack(alignment: .leading, spacing: 0) {
-            ForEach(Array(completion.candidates.enumerated()), id: \.offset) { index, candidate in
-                HStack(spacing: 6) {
-                    Image(systemName: isSlash ? "command" : "doc")
-                        .font(.system(size: 10))
-                        .foregroundStyle(theme.agentTextFg.opacity(0.4))
-                        .frame(width: 14)
-
-                    Text(candidate.name)
-                        .font(.system(size: 12, design: .monospaced))
-                        .foregroundStyle(theme.agentTextFg)
-                        .lineLimit(1)
-
-                    if !candidate.description.isEmpty {
-                        Text(candidate.description)
-                            .font(.system(size: 11))
-                            .foregroundStyle(theme.agentTextFg.opacity(0.4))
-                            .lineLimit(1)
-                    }
-
-                    Spacer()
-                }
-                .padding(.horizontal, 8)
-                .padding(.vertical, 4)
-                .background(index == Int(completion.selected) ? theme.agentInputBorder.opacity(0.15) : Color.clear)
-            }
-        }
-        .frame(maxWidth: 400)
-        .frame(maxHeight: CGFloat(completion.candidates.count) * 24)
-        .background(
-            RoundedRectangle(cornerRadius: 6)
-                .fill(theme.agentCodeBg)
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 6)
-                .strokeBorder(theme.agentCodeBorder.opacity(0.4), lineWidth: 1)
-        )
-        .shadow(color: .black.opacity(0.2), radius: 8, y: -4)
-        .padding(.horizontal, 16)
-        .accessibilityElement(children: .contain)
-        .accessibilityLabel(isSlash ? "Slash command completion" : "File mention completion")
     }
 
     // MARK: - Help overlay
@@ -1212,71 +546,5 @@ struct AgentChatView: View {
         case (_, .user): return true
         default: return false
         }
-    }
-
-    // MARK: - Helpers
-
-    private var statusColor: Color {
-        switch state.status {
-        case 0: return Color.gray
-        case 1: return theme.agentHeaderFg
-        case 2: return Color.orange
-        case 3: return Color.red
-        default: return Color.gray
-        }
-    }
-
-    private func toolIcon(_ status: UInt8) -> String {
-        switch status {
-        case 0: return "gearshape"
-        case 1: return "checkmark.circle"
-        case 2: return "exclamationmark.triangle"
-        default: return "gearshape"
-        }
-    }
-
-    @ViewBuilder
-    private func statusBadge(_ status: UInt8, isError: Bool) -> some View {
-        let (text, color): (String, Color) = {
-            if isError { return ("error", Color.red) }
-            switch status {
-            case 0: return ("running", Color.orange)
-            case 1: return ("done", Color.green)
-            case 2: return ("error", Color.red)
-            default: return ("", Color.clear)
-            }
-        }()
-
-        if !text.isEmpty {
-            Text(text)
-                .font(.system(size: 9, weight: .medium))
-                .foregroundStyle(color.opacity(0.8))
-                .padding(.horizontal, 6)
-                .padding(.vertical, 2)
-                .background(
-                    RoundedRectangle(cornerRadius: 4)
-                        .fill(color.opacity(0.1))
-                )
-        }
-    }
-
-    private func autoApprovedPill(label: String) -> some View {
-        Text("auto-approved · \(label)")
-            .font(.system(size: 10, design: .rounded))
-            .foregroundStyle(theme.agentTextFg.opacity(0.55))
-            .accessibilityLabel("Auto-approved for this \(label)")
-    }
-
-    private func autoApprovedScopeLabel(_ scope: UInt8) -> String? {
-        switch scope {
-        case 1: return "session"
-        case 2: return "turn"
-        default: return nil
-        }
-    }
-
-    private func formatDuration(_ ms: UInt32) -> String {
-        if ms < 1000 { return "\(ms)ms" }
-        return String(format: "%.1fs", Double(ms) / 1000.0)
     }
 }
